chore: refactor babylond with better DevEx
