+++ modules/babylond/chainio.go
@@ -10,13 +10,17 @@ import (
 	transactionprocess "github.com/satlayer/satlayer-bvs/bvs-api/metrics/indicators/transaction_process"
 )
 
-func (c *BabylonContainer) NewChainIO(homeDir string) (io.ChainIO, error) {
+func (c *BabylonContainer) NewChainIO(homeDir string) io.ChainIO {
 	logger := apilogger.NewMockELKLogger()
-	metricsIndicators := transactionprocess.NewPromIndicators(prometheus.NewRegistry(), "io")
-	return io.NewChainIO(ChainId, c.RpcUri, homeDir, "bbn", logger, metricsIndicators, types.TxManagerParams{
+	metricsIndicators := transactionprocess.NewPromIndicators(prometheus.NewRegistry(), "testcontainers")
+	chainIo, err := io.NewChainIO(ChainId, c.RpcUri, homeDir, "bbn", logger, metricsIndicators, types.TxManagerParams{
 		MaxRetries:             3,
 		RetryInterval:          2 * time.Second,
 		ConfirmationTimeout:    60 * time.Second,
 		GasPriceAdjustmentRate: "1.1",
 	})
+	if err != nil {
+		panic(err)
+	}
+	return chainIo
 }
